+++ WearComplicationProvidersTestSuite/Wearable/src/main/java/com/example/android/wearable/wear/wearcomplicationproviderstestsuite/NoDataDataSourceService.kt
@@ -24,7 +24,7 @@ import androidx.wear.complications.datasource.ComplicationRequest
 /**
  * A complication provider that always returns [ComplicationType.NO_DATA].
  */
-class NoDataProviderService : ComplicationDataSourceService() {
+class NoDataDataSourceService : ComplicationDataSourceService() {
     override fun onComplicationRequest(
         request: ComplicationRequest,
         listener: ComplicationRequestListener
