+++ WearComplicationProvidersTestSuite/Wearable/src/main/java/com/example/android/wearable/wear/wearcomplicationproviderstestsuite/ShortTextDataSourceService.kt
@@ -22,14 +22,14 @@ import androidx.datastore.core.DataStore
 import androidx.wear.complications.data.ComplicationData
 import androidx.wear.complications.data.ComplicationType
 import androidx.wear.complications.data.MonochromaticImage
-import androidx.wear.complications.data.MonochromaticImageComplicationData
 import androidx.wear.complications.data.PlainComplicationText
+import androidx.wear.complications.data.ShortTextComplicationData
 import androidx.wear.complications.datasource.ComplicationDataSourceService
 import androidx.wear.complications.datasource.ComplicationRequest
 
 /**
- * A complication provider that supports only [ComplicationType.MONOCHROMATIC_IMAGE] and cycles through
- * a few different icons on each tap.
+ * A complication provider that supports only [ComplicationType.SHORT_TEXT] and cycles
+ * through the possible configurations on tap.
  *
  * Note: This subclasses [SuspendingComplicationDataSourceService] instead of [ComplicationDataSourceService] to support
  * coroutines, so data operations (specifically, calls to [DataStore]) can be supported directly in the
@@ -38,17 +38,16 @@ import androidx.wear.complications.datasource.ComplicationRequest
  *
  * If you don't perform any suspending operations to update your complications, you can subclass
  * [ComplicationDataSourceService] and override [onComplicationRequest] directly.
- * (see [NoDataProviderService] for an example)
+ * (see [NoDataDataSourceService] for an example)
  */
-class IconProviderService : SuspendingComplicationDataSourceService() {
-
+class ShortTextDataSourceService : SuspendingComplicationDataSourceService() {
     override suspend fun onComplicationRequest(request: ComplicationRequest): ComplicationData? {
-        if (request.complicationType != ComplicationType.MONOCHROMATIC_IMAGE) {
+        if (request.complicationType != ComplicationType.SHORT_TEXT) {
             return null
         }
         val args = ComplicationToggleArgs(
             providerComponent = ComponentName(this, javaClass),
-            complication = Complication.ICON,
+            complication = Complication.SHORT_TEXT,
             complicationInstanceId = request.complicationInstanceId
         )
         val complicationTogglePendingIntent =
@@ -57,7 +56,7 @@ class IconProviderService : SuspendingComplicationDataSourceService() {
                 args = args
             )
         // Suspending function to retrieve the complication's state
-        val state = args.getState(this@IconProviderService)
+        val state = args.getState(this)
         val case = Case.values()[state.mod(Case.values().size)]
         return getComplicationData(
             tapAction = complicationTogglePendingIntent,
@@ -68,45 +67,76 @@ class IconProviderService : SuspendingComplicationDataSourceService() {
     override fun getPreviewData(type: ComplicationType): ComplicationData =
         getComplicationData(
             tapAction = null,
-            case = Case.FACE,
+            case = Case.TEXT_WITH_ICON_AND_TITLE
         )
 
     private fun getComplicationData(
         tapAction: PendingIntent?,
-        case: Case,
+        case: Case
     ): ComplicationData =
         when (case) {
-            Case.FACE -> MonochromaticImageComplicationData.Builder(
-                monochromaticImage = MonochromaticImage.Builder(
-                    Icon.createWithResource(this, R.drawable.ic_face_vd_theme_24)
+            Case.TEXT_ONLY -> ShortTextComplicationData.Builder(
+                text = PlainComplicationText.Builder(
+                    text = getText(R.string.short_text_only)
                 ).build(),
                 contentDescription = PlainComplicationText.Builder(
-                    text = getText(R.string.icon_face_content_description)
+                    text = getText(R.string.short_text_only_content_description)
                 ).build()
             )
-            Case.BATTERY -> MonochromaticImageComplicationData.Builder(
-                monochromaticImage = MonochromaticImage.Builder(
-                    Icon.createWithResource(this, R.drawable.ic_battery)
+            Case.TEXT_WITH_ICON -> ShortTextComplicationData.Builder(
+                text = PlainComplicationText.Builder(
+                    text = getText(R.string.short_text_with_icon)
+                ).build(),
+                contentDescription = PlainComplicationText.Builder(
+                    text = getText(R.string.short_text_with_icon_content_description)
+                ).build()
+            )
+                .setMonochromaticImage(
+                    MonochromaticImage.Builder(
+                        image = Icon.createWithResource(this, R.drawable.ic_face_vd_theme_24)
+                    ).build()
                 )
-                    .setAmbientImage(Icon.createWithResource(this, R.drawable.ic_battery_burn_protect))
-                    .build(),
-                contentDescription = PlainComplicationText.Builder(
-                    text = getText(R.string.icon_battery_content_description)
-                ).build()
-            )
-            Case.EVENT -> MonochromaticImageComplicationData.Builder(
-                monochromaticImage = MonochromaticImage.Builder(
-                    Icon.createWithResource(this, R.drawable.ic_event_vd_theme_24)
+            Case.TEXT_WITH_TITLE -> ShortTextComplicationData.Builder(
+                text = PlainComplicationText.Builder(
+                    text = getText(R.string.short_text_with_title)
                 ).build(),
                 contentDescription = PlainComplicationText.Builder(
-                    text = getText(R.string.icon_event_content_description)
+                    text = getText(R.string.short_text_with_title_content_description)
                 ).build()
             )
+                .setTitle(
+                    PlainComplicationText.Builder(
+                        text = getText(R.string.short_title)
+                    ).build()
+                )
+            Case.TEXT_WITH_ICON_AND_TITLE -> ShortTextComplicationData.Builder(
+                // When short text includes both short title and icon, the watch face should only
+                // display one of those fields.
+                text = PlainComplicationText.Builder(
+                    text = getText(R.string.short_text_with_both)
+                ).build(),
+                contentDescription = PlainComplicationText.Builder(
+                    text = getText(R.string.short_text_with_both_content_description)
+                ).build()
+            )
+                .setTitle(
+                    PlainComplicationText.Builder(
+                        text = getText(R.string.short_title)
+                    ).build()
+                )
+                .setMonochromaticImage(
+                    MonochromaticImage.Builder(
+                        image = Icon.createWithResource(this, R.drawable.ic_face_vd_theme_24)
+                    ).build()
+                )
         }
             .setTapAction(tapAction)
             .build()
 
     private enum class Case {
-        FACE, BATTERY, EVENT
+        TEXT_ONLY,
+        TEXT_WITH_ICON,
+        TEXT_WITH_TITLE,
+        TEXT_WITH_ICON_AND_TITLE
     }
 }
